Move DocumentCacheService
